requirements.txt
numpy
matplotlib
pyprojroot
pydrive
tifffile
scipy
gdown
tqdm
opencv-python
umap-learn
scikit-learn
torch
torchvision
wandb
torchmetrics
pytorch-lightning
segmentation-models-pytorch
rich

